rename /opengate/helpers.py to utility.py
