+++ opengate/actors/filter_builders.py
@@ -1,5 +1,5 @@
 from .filters import KineticEnergyFilter, ParticleFilter, TrackCreatorProcessFilter
-from ..helpers import make_builders
+from ..utility import make_builders
 
 filter_type_names = {ParticleFilter, KineticEnergyFilter, TrackCreatorProcessFilter}
 filter_builders = make_builders(filter_type_names)
