+++ opengate/bin/opengate_info
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-from opengate.helpers import print_opengate_info
+from opengate.utility import print_opengate_info
 
 if __name__ == "__main__":
     print_opengate_info()
