+++ opengate/contrib/dose/doserate.py
@@ -3,7 +3,7 @@
 
 import pathlib
 from opengate.managers import Simulation
-from opengate.helpers import g4_units, g4_best_unit
+from opengate.utility import g4_units, g4_best_unit
 from opengate.image import get_translation_between_images_center, read_image_info
 from opengate.logger import INFO
 from opengate.geometry.materials import HounsfieldUnit_to_material
