+++ opengate/contrib/pet/philipsvereos.py
@@ -1,5 +1,5 @@
 from scipy.spatial.transform import Rotation
-from opengate.helpers import g4_units
+from opengate.utility import g4_units
 from opengate.geometry.utility import repeat_ring, repeat_array
 
 # colors (similar to the ones of Gate)
