+++ opengate/contrib/spect/genm670.py
@@ -1,6 +1,6 @@
 import pathlib
 from opengate.exception import fatal
-from opengate.helpers import g4_units
+from opengate.utility import g4_units
 
 from opengate.geometry.utility import (
     get_volume_bounding_limits,
