+++ opengate/geometry/ImageVolume.py
@@ -2,7 +2,7 @@ import itk
 import numpy as np
 import opengate_core as g4
 from .VolumeBase import VolumeBase
-from ..helpers import check_filename_type
+from ..utility import check_filename_type
 from ..image import read_image_info, create_3d_image, update_image_py_to_cpp
 from .utility import get_vol_g4_transform
