+++ opengate/runtiming.py
@@ -1,4 +1,4 @@
-from .helpers import g4_best_unit, indent
+from .utility import g4_best_unit, indent
 from .exception import fatal
