+++ opengate/tests/src/test003_G4Material.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 import opengate as gate
-from opengate.helpers import g4_units
+from opengate.utility import g4_units
 import opengate.tests.utility as utility
 import opengate_core as g4
 import math
