+++ opengate/tests/src/test004_simple_MT.py
@@ -3,7 +3,7 @@
 
 import opengate as gate
 import opengate.tests.utility as utility
-from opengate.helpers import g4_units
+from opengate.utility import g4_units
 
 if __name__ == "__main__":
     paths = utility.get_default_test_paths(
