+++ opengate/utility.py
